[fix] Sound for main menu
