+++ Assets/Scripts/Scenes/CreditsController.cs
